feat: Use yelp review_count instead of is_closed
Earlier versions displayed if a location was currently open or
closed, based on the `is_closed` attribute of the object returned
from the Yelp API.  Unfortunately, that attribute is about whether
the business is permanently closed or not; not about business
hours.  So I had to use a different attribute.  I chose the review
count.