bugfix: Do not reselect selectedPlace, only bounce
If the user clicks on a Place in the search list that is already
selected, that Place's map marker will bounce, but nothing else
will occur.  This prevents a bug when the yelp error message is
showing and the user clicks the selectedPlace from the search
list when back online.  This would cause the yelpData to fill
the Info Window along with the error message.  This bug is fixed
now.