bugfix: Stop multiple IWs from being open at once
Prior to this commit, if the user selected a Place, then filtered
the Place with the search input, then cleared the search input and
selected another Place, the original info window would not close
and the info window for the second Place would open, leaving more
than one info window open at one time.  This was BAD!  This has
been fixed.  The problem was that, if the search filtered the
selectedPlace, it would deselect that place.  This no longer
occurs and the bug has been fixed!