refactor: Fill IWs with knockout instead of jquery
My project reviewer pointed out that I need to use knockout to
fill in the data in the info windows, instead of using jquery to
do so.  The app seems to be working correctly, but there is still
a fair bit of code clean up to do before I can resubmit.