bugfix: Searching deselects Place if no match
Prior to this commit, if an info window was open and the user
entered a value into the search that did not match the selected
place, the map marker for the selected place would disappear but
the info window would stay visible.  Now, searching can deselect
the selectedPlace if the search doesn't match the name of the
selected Place.