bugfix: Reset ajaxErrorOccurred attr on iw close
There was a bug occurring where if an info window was opened
and an ajax error occurred, and then the IW was closed and
reopened without an ajax error occurring, the error message would
still be in the info window below the yelp data.

This was caused by the ajaxErrorOccurred Place attribute not
getting reset to false after an info window was closed.
The attribute now gets reset whenever the info window closes.